[IMP] l10n_my_edi*: MyInvois Documents in accounting
When recently implementing support for the PoS,
we started to plan an improvement of the whole
feature by implementing a MyInvois Document model
to regroup the flows in one place.

These documents aim to be the representation of a
document on the myinvois platform, and can be backed
by whichever Odoo model needs to be sent to MyInvois
(so far, PoS orders and invoices).

So far, invoices have not been linked to a MyInvois
Document, meaning that the code we have twice the code
to send and manage documents in the code base.
Once for MyInvois Documents, once for invoices.

In this PR, we are linking the invoices to this
Document Model, allowing to remove the duplicated code
on the account.move model.

task-4929885

closes odoo/odoo#220034

Related: odoo/upgrade#8107
