+++ addons/l10n_my_edi/__manifest__.py
@@ -16,10 +16,12 @@
         'data/l10n_my_edi.industry_classification.csv',
 
         'security/ir.model.access.csv',
+        'security/myinvois_security.xml',
 
         'views/account_move_view.xml',
         'views/account_tax_view.xml',
         'views/l10n_my_edi_industrial_classification_views.xml',
+        "views/myinvois_document_views.xml",
         'views/product_template_view.xml',
         'views/report_invoice.xml',
         'views/res_company_view.xml',
@@ -27,7 +29,8 @@
         'views/res_partner_view.xml',
         'views/account_portal_templates.xml',
 
-        'wizard/l10n_my_edi_status_update_wizard.xml',
+        'wizard/myinvois_consolidate_invoice_wizard.xml',
+        'wizard/myinvois_document_status_update_wizard.xml',
     ],
     'installable': True,
     'author': 'Odoo S.A.',
