+++ addons/l10n_my_edi/data/ir_cron.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" ?>
 <odoo>
     <data>
-        <record id="ir_cron_myinvois_sync" model="ir.cron">
-            <field name="name">MyInvois: Synchronization</field>
-            <field name="model_id" ref="account.model_account_move"/>
+        <record id="ir_cron_myinvois_document_sync" model="ir.cron">
+            <field name="name">MyInvois: Document Synchronization</field>
+            <field name="model_id" ref="l10n_my_edi.model_myinvois_document"/>
             <field name="state">code</field>
-            <field name="code">model._cron_l10n_my_edi_synchronize_myinvois()</field>
+            <field name="code">model._myinvois_statuses_update_cron()</field>
             <field name="interval_number">1</field>
             <field name="interval_type">hours</field>
             <field name="nextcall" eval="(DateTime.now() + timedelta(hours=1))"/>
