+++ addons/l10n_my_edi/data/l10n_my_edi.industry_classification.csv
@@ -279,8 +279,9 @@
 "class_15202","15202","Manufacture of plastic footwear"
 "class_15203","15203","Manufacture of rubber footwear"
 "class_15209","15209","Manufacture of other footwear n.e.c."
+"class_16100","16100","Sawmilling and planning of wood"
 "class_16211","16211","Manufacture of veneer sheets and plywood"
-"class_16221","16221","Manufacture of builders' carpentry"
+"class_16212","16212","Manufacture of particle board and fibreboard"
 "class_16222","16222","Manufacture of joinery wood products"
 "class_16230","16230","Manufacture of wooden containers"
 "class_16291","16291","Manufacture of wood charcoal"
@@ -673,10 +674,12 @@
 "class_47222","47222","Retail sale of tea, coffee, soft drinks, mineral water and other beverages"
 "class_47230","47230","Retail sale of tobacco products in specialized store"
 "class_47300","47300","Retail sale of automotive fuel in specialized stores"
+"class_47411","47411","Retail sale of computers, computer equipment and supplies"
 "class_47412","47412","Retail sale of video game consoles and non-customized software"
 "class_47413","47413","Retail sale of telecommunication equipment"
 "class_47420","47420","Retail sale of audio and video equipment in specialized store"
 "class_47510","47510","Retail sale of textiles in specialized stores"
+"class_47520","47520","Retail sale of construction materials, hardware, paints and glass"
 "class_47531","47531","Retail sale of carpets and rugs"
 "class_47532","47532","Retail sale of curtains and net curtains"
 "class_47533","47533","Retail sale of wallpaper and floor coverings"
@@ -689,6 +692,7 @@
 "class_47597","47597","Retail sale of security systems"
 "class_47598","47598","Retail sale of household articles and equipment n.e.c."
 "class_47611","47611","Retail sale of office supplies and equipment"
+"class_47620","47620","Retail sale of musical records, audio tapes, compact discs , cassettes, video tapes, VCDs and DVDs, blank tapes and discs"
 "class_47612","47612","Retail sale of books, newspapers and stationary"
 "class_47631","47631","Retail sale of sports goods and equipments"
 "class_47632","47632","Retail sale of fishing equipment"
@@ -717,7 +721,8 @@
 "class_47749","47749","Retail sale of second-hand goods n.e.c."
 "class_47810","47810","Retail sale of food, beverages and tobacco products via stalls or markets"
 "class_47820","47820","Retail sale of textiles, clothing and footwear via stalls or markets"
-"class_47891","47891","Retail sale of carpets and rugs via stalls or markets"
+"class_47891","47891","Retail sale of books via stalls or markets"
+"class_47892","47892","Retail sale of carpets and rugs via stalls or markets"
 "class_47893","47893","Retail sale of games and toys via stalls or markets"
 "class_47894","47894","Retail sale of household appliances and consumer electronics via stall or markets"
 "class_47895","47895","Retail sale of music and video recordings via stall or markets"
@@ -725,6 +730,7 @@
 "class_47912","47912","Retail sale of any kind of product over the Internet"
 "class_47913","47913","Direct sale via television, radio and telephone"
 "class_47914","47914","Internet retail auctions"
+"class_47991","47991","Retail sale of any kind of product by direct sales or door-to-door sales persons"
 "class_47992","47992","Retail sale of any kind of product through vending machines"
 "class_47999","47999","Other retail sale not in stores, stalls or markets n.e.c."
 "class_49110","49110","Passenger transport by inter-urban railways"
@@ -785,6 +791,8 @@
 "class_55109","55109","Other short term accommodation activities n.e.c."
 "class_55200","55200","Camping grounds, recreational vehicle parks and trailer parks"
 "class_55900","55900","Other accommodation"
+"class_56101","56101","Restaurants and restaurant cum night clubs"
+"class_56102","56102","Cafeterias/canteens "
 "class_56103","56103","Fast-food restaurants"
 "class_56104","56104","Ice cream truck vendors and parlours"
 "class_56105","56105","Mobile food carts"
