+++ addons/l10n_my_edi/models/__init__.py
@@ -6,6 +6,7 @@ from . import account_move_line
 from . import account_move_send
 from . import account_tax
 from . import l10n_my_edi_industry_classification
+from . import myinvois_document
 from . import product_template
 from . import res_company
 from . import res_config_settings
