+++ addons/l10n_my_edi/models/account_move_send.py
@@ -1,5 +1,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from odoo import api, models
 
 
@@ -16,5 +15,5 @@ class AccountMoveSend(models.AbstractModel):
         # EXTENDS 'account'
         return (
             super()._get_invoice_extra_attachments(move)
-            + move.l10n_my_edi_file_id
+            + move._get_active_myinvois_document().myinvois_file_id
         )
