+++ addons/l10n_my_edi/models/account_tax.py
@@ -1,5 +1,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from odoo import api, fields, models
 
 
@@ -24,6 +23,10 @@ class AccountTax(models.Model):
         compute="_compute_l10n_my_tax_type",
         store=True,
         readonly=False,
+    )
+    l10n_my_tax_exemption_reason = fields.Char(
+        string="Malaysian Tax Exemption Reason",
+        help="The reason for tax exemption, used when submitting consolidated invoices including this tax.",
     )
 
     # --------------------------------
