+++ addons/l10n_my_edi/models/l10n_my_edi_industry_classification.py
@@ -1,5 +1,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from odoo import api, fields, models
