+++ addons/l10n_my_edi/models/product_template.py
@@ -1,5 +1,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from odoo import fields, models
 
 CLASSIFICATION_CODES_LIST = [
