+++ addons/l10n_my_edi/security/ir.model.access.csv
@@ -2,6 +2,7 @@
 access_l10n_my_edi_industry_classification_readonly,l10n_my_edi.industry_classification,model_l10n_my_edi_industry_classification,account.group_account_readonly,1,0,0,0
 access_l10n_my_edi_industry_classification_invoice,l10n_my_edi.industry_classification,model_l10n_my_edi_industry_classification,account.group_account_invoice,1,0,0,0
 access_l10n_my_edi_industry_classification_manager,l10n_my_edi.industry_classification,model_l10n_my_edi_industry_classification,account.group_account_manager,1,1,1,1
-access_l10n_my_edi_document_status_update_readonly,l10n_my_edi.document.status.update,model_l10n_my_edi_document_status_update,account.group_account_readonly,1,0,0,0
-access_l10n_my_edi_document_status_update_invoice,l10n_my_edi.document.status.update,model_l10n_my_edi_document_status_update,account.group_account_invoice,1,0,0,0
-access_l10n_my_edi_document_status_update_manager,l10n_my_edi.document.status.update,model_l10n_my_edi_document_status_update,account.group_account_manager,1,1,1,1
+access_myinvois_consolidate_invoice_wizard_user,access_myinvois_consolidate_invoice_wizard_user,model_myinvois_consolidate_invoice_wizard,account.group_account_invoice,1,1,1,1
+access_myinvois_document_status_update_wizard_invoice,l10n_my_edi.myinvois.document.status.update.wizard,model_myinvois_document_status_update_wizard,account.group_account_invoice,1,1,1,1
+access_myinvois_document_readonly,l10n_my_edi.myinvois.document,model_myinvois_document,account.group_account_readonly,1,0,0,0
+access_myinvois_document_invoice,l10n_my_edi.myinvois.document,model_myinvois_document,account.group_account_invoice,1,1,1,1
