+++ addons/l10n_my_edi/tests/expected_xmls/bill_import.xml
@@ -5,7 +5,6 @@
   <cbc:IssueTime>10:00:00Z</cbc:IssueTime>
   <cbc:InvoiceTypeCode listVersionID="1.1">11</cbc:InvoiceTypeCode>
   <cbc:DocumentCurrencyCode>MYR</cbc:DocumentCurrencyCode>
-  <cbc:BuyerReference>MY-REF</cbc:BuyerReference>
   <cac:AdditionalDocumentReference>
     <cbc:ID>E12345678912</cbc:ID>
     <cbc:DocumentType>CustomsImportForm</cbc:DocumentType>
