+++ addons/l10n_my_edi/tests/expected_xmls/credit_note.xml
@@ -104,6 +104,9 @@
       </cac:PartyLegalEntity>
     </cac:DeliveryParty>
   </cac:Delivery>
+  <cac:PrepaidPayment>
+      <cbc:PaidAmount currencyID="EUR">2200.00</cbc:PaidAmount>
+  </cac:PrepaidPayment>
   <cac:TaxExchangeRate>
     <cbc:SourceCurrencyCode>EUR</cbc:SourceCurrencyCode>
     <cbc:TargetCurrencyCode>MYR</cbc:TargetCurrencyCode>
@@ -128,7 +131,7 @@
     <cbc:LineExtensionAmount currencyID="EUR">2000.00</cbc:LineExtensionAmount>
     <cbc:TaxExclusiveAmount currencyID="EUR">2000.00</cbc:TaxExclusiveAmount>
     <cbc:TaxInclusiveAmount currencyID="EUR">2200.00</cbc:TaxInclusiveAmount>
-    <cbc:PrepaidAmount currencyID="EUR">2200.00</cbc:PrepaidAmount>
+    <cbc:PrepaidAmount currencyID="EUR">0.00</cbc:PrepaidAmount>
     <cbc:PayableAmount currencyID="EUR">0.00</cbc:PayableAmount>
   </cac:LegalMonetaryTotal>
   <cac:InvoiceLine>
