+++ addons/l10n_my_edi/tests/expected_xmls/refund_note_company_currency.xml
@@ -103,6 +103,9 @@
       </cac:PartyLegalEntity>
     </cac:DeliveryParty>
   </cac:Delivery>
+  <cac:PrepaidPayment>
+      <cbc:PaidAmount currencyID="MYR">1100.00</cbc:PaidAmount>
+  </cac:PrepaidPayment>
   <cac:TaxTotal>
     <cbc:TaxAmount currencyID="MYR">100.00</cbc:TaxAmount>
     <cac:TaxSubtotal>
@@ -122,7 +125,7 @@
     <cbc:LineExtensionAmount currencyID="MYR">1000.00</cbc:LineExtensionAmount>
     <cbc:TaxExclusiveAmount currencyID="MYR">1000.00</cbc:TaxExclusiveAmount>
     <cbc:TaxInclusiveAmount currencyID="MYR">1100.00</cbc:TaxInclusiveAmount>
-    <cbc:PrepaidAmount currencyID="MYR">1100.00</cbc:PrepaidAmount>
+    <cbc:PrepaidAmount currencyID="MYR">0.00</cbc:PrepaidAmount>
     <cbc:PayableAmount currencyID="MYR">0.00</cbc:PayableAmount>
   </cac:LegalMonetaryTotal>
   <cac:InvoiceLine>
