+++ addons/l10n_my_edi/views/account_tax_view.xml
@@ -7,6 +7,7 @@
         <field name="arch" type="xml">
             <xpath expr="//field[@name='active']" position="after">
                 <field name="l10n_my_tax_type" invisible="country_code != 'MY'"/>
+                <field name="l10n_my_tax_exemption_reason" invisible="l10n_my_tax_type != 'E'"/>
             </xpath>
         </field>
     </record>
