+++ addons/l10n_my_edi/views/report_invoice.xml
@@ -2,7 +2,7 @@
 <odoo>
     <template id="report_invoice_document" inherit_id="account.report_invoice_document" primary="True">
         <div id="qrcode" position="after">
-            <div id="myinvois_qrcode" class="d-flex mb-3 avoid-page-break-inside" t-if="o.l10n_my_edi_invoice_long_id">
+            <div id="myinvois_qrcode" class="d-flex mb-3 avoid-page-break-inside" t-if="o._get_active_myinvois_document().myinvois_document_long_id">
                 <div class="qrcode me-3" id="myinvois_qrcode_image">
                     <t t-set="qr_code_url" t-value="o._generate_myinvois_qr_code()"/>
                     <p t-if="qr_code_url" class="position-relative mb-0">
