+++ addons/l10n_my_edi/wizard/__init__.py
@@ -1,2 +1,3 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-from . import l10n_my_edi_status_update_wizard
+from . import myinvois_consolidate_invoice_wizard
+from . import myinvois_document_status_update_wizard
