--- addons/l10n_my_edi/wizard/l10n_my_edi_status_update_wizard.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="l10n_my_edi_document_status_update_form" model="ir.ui.view">
-        <field name="name">l10n_my_edi.document.status.update.form</field>
-        <field name="model">l10n_my_edi.document.status.update</field>
-        <field name="arch" type="xml">
-            <form>
-                <group>
-                    <field name="invoice_id"/>
-                    <field name="new_status"/>
-                    <field name="reason"/>
-                </group>
-                <footer>
-                    <button string="Update Invoice" name="button_request_update" type="object" default_focus="1" class="btn-primary"/>
-                    <button string="Close" special="cancel"/>
-                </footer>
-            </form>
-        </field>
-    </record>
-</odoo>
