+++ addons/l10n_my_edi/wizard/myinvois_consolidate_invoice_wizard.xml
@@ -14,7 +14,7 @@
                     </group>
                 </sheet>
                 <footer>
-                    <button string="Create Consolidated Invoices" name="button_consolidate_orders" type="object" default_focus="1" class="btn-primary"/>
+                    <button string="Create Consolidated Invoices" name="button_consolidate" type="object" default_focus="1" class="btn-primary"/>
                     <button string="Close" special="cancel"/>
                 </footer>
             </form>
