+++ addons/l10n_my_edi_pos/__manifest__.py
@@ -11,20 +11,11 @@
     """,
     "depends": ["l10n_my_edi", "point_of_sale"],
     "data": [
-        "data/ir_cron.xml",
         "data/res_partner.xml",
 
-        "security/myinvois_security.xml",
-        "security/ir.model.access.csv",
-
-        "views/account_tax_view.xml",
-        "views/myinvois_document_views.xml",
         "views/myinvois_document_pos_views.xml",
         "views/pos_order_views.xml",
         "views/product_view.xml",
-
-        "wizard/myinvois_consolidate_invoice_wizard.xml",
-        "wizard/myinvois_status_update_wizard.xml",
     ],
     "installable": True,
     'auto_install': True,
