+++ addons/l10n_my_edi_pos/i18n/l10n_my_edi_pos.pot
@@ -1,0 +1,160 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* l10n_my_edi_pos
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 18.5a1+e\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2025-07-23 07:08+0000\n"
+"PO-Revision-Date: 2025-07-23 07:08+0000\n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: l10n_my_edi_pos
+#: model_terms:ir.ui.view,arch_db:l10n_my_edi_pos.view_pos_pos_form
+msgid "<span>Consolidated<br/>Invoice</span>"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model,name:l10n_my_edi_pos.model_myinvois_consolidate_invoice_wizard
+msgid "Consolidate Invoice Wizard"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model_terms:ir.ui.view,arch_db:l10n_my_edi_pos.myinvois_document_pos_list_view
+msgid "Consolidate Orders"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_pos_order__consolidated_invoice_ids
+#: model:ir.ui.menu,name:l10n_my_edi_pos.menu_consolidated_invoices
+msgid "Consolidated Invoice"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#. odoo-python
+#: code:addons/l10n_my_edi_pos/models/myinvois_document_pos.py:0
+#: code:addons/l10n_my_edi_pos/models/pos_order.py:0
+#: model:ir.actions.act_window,name:l10n_my_edi_pos.action_consolidated_invoices
+msgid "Consolidated Invoices"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_myinvois_consolidate_invoice_wizard__consolidation_type
+msgid "Consolidation Type"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#. odoo-python
+#: code:addons/l10n_my_edi_pos/models/myinvois_document_pos.py:0
+msgid "Create Consolidated Invoice"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_myinvois_document__pos_order_date_range
+msgid "Date Range"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_account_edi_xml_ubl_myinvois_my__display_name
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_myinvois_consolidate_invoice_wizard__display_name
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_myinvois_document__display_name
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_pos_order__display_name
+msgid "Display Name"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_account_edi_xml_ubl_myinvois_my__id
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_myinvois_consolidate_invoice_wizard__id
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_myinvois_document__id
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_pos_order__id
+msgid "ID"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#. odoo-python
+#: code:addons/l10n_my_edi_pos/wizard/myinvois_consolidate_invoice_wizard.py:0
+msgid "Invalid Operation. No order to consolidate."
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_myinvois_document__linked_order_count
+msgid "Linked Order Count"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model,name:l10n_my_edi_pos.model_account_edi_xml_ubl_myinvois_my
+msgid "Malaysian implementation of ubl for the MyInvois portal"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model,name:l10n_my_edi_pos.model_myinvois_document
+msgid "MyInvois Document"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_myinvois_document__pos_config_id
+msgid "PoS Config"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model.fields.selection,name:l10n_my_edi_pos.selection__myinvois_consolidate_invoice_wizard__consolidation_type__pos
+msgid "PoS Order"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model_terms:ir.ui.view,arch_db:l10n_my_edi_pos.myinvois_document_pos_form_view
+msgid "PoS Orders"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#. odoo-python
+#: code:addons/l10n_my_edi_pos/models/myinvois_document_pos.py:0
+#: model:ir.model,name:l10n_my_edi_pos.model_pos_order
+msgid "Point of Sale Orders"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#: model:ir.model.fields,field_description:l10n_my_edi_pos.field_myinvois_document__pos_order_ids
+msgid "Pos Order"
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#. odoo-python
+#: code:addons/l10n_my_edi_pos/models/pos_order.py:0
+msgid ""
+"This order has been included in a consolidated invoice and cannot be "
+"invoiced separately."
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#. odoo-python
+#: code:addons/l10n_my_edi_pos/models/pos_order.py:0
+msgid ""
+"You cannot invoice a refund for an order that has not been submitted to "
+"MyInvois yet."
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#. odoo-python
+#: code:addons/l10n_my_edi_pos/models/pos_order.py:0
+msgid ""
+"You must invoice a refund for an order that has been submitted to MyInvois."
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#. odoo-python
+#: code:addons/l10n_my_edi_pos/models/pos_order.py:0
+msgid "You must set a TIN number on the commercial partner."
+msgstr ""
+
+#. module: l10n_my_edi_pos
+#. odoo-python
+#: code:addons/l10n_my_edi_pos/models/pos_order.py:0
+msgid "You must set the identification information on the commercial partner."
+msgstr ""
