+++ addons/l10n_my_edi_pos/models/pos_order.py
@@ -95,7 +95,7 @@ class PosOrder(models.Model):
 
             # At this point we don't want to raise anymore, if there are issues it'll be logged on the invoice, and we will
             # move on.
-            self.account_move._l10n_my_edi_send_invoice(commit=False)
+            self.account_move.action_l10n_my_edi_send_invoice()
 
             if self.env.context.get('generate_pdf', True):
                 self.account_move.with_context(skip_invoice_sync=True)._generate_and_send()
@@ -107,30 +107,13 @@ class PosOrder(models.Model):
     # Action methods
     # --------------
 
-    def action_show_consolidated_invoice(self):
-        if len(self._get_active_consolidated_invoice()) == 1:
-            action_vals = {
-                'type': 'ir.actions.act_window',
-                'res_model': 'myinvois.document',
-                'view_mode': 'form',
-                'res_id': self._get_active_consolidated_invoice().id,
-                'views': [(self.env.ref('l10n_my_edi_pos.myinvois_document_pos_form_view').id, 'form')],
-            }
-        else:
-            action_vals = {
-                'name': self.env._("Consolidated Invoices"),
-                'type': 'ir.actions.act_window',
-                'res_model': 'myinvois.document',
-                'view_mode': 'list,form',
-                'views': [(self.env.ref('l10n_my_edi_pos.myinvois_document_pos_list_view').id, 'list'), (self.env.ref('l10n_my_edi_pos.myinvois_document_pos_form_view').id, 'form')],
-                'domain': [('id', 'in', self._get_active_consolidated_invoice().ids)],
-            }
-        return action_vals
+    def action_show_myinvois_documents(self):
+        return self._get_active_consolidated_invoice().action_show_myinvois_documents()
 
     # ----------------
     # Business methods
     # ----------------
 
-    def _get_active_consolidated_invoice(self):
+    def _get_active_consolidated_invoice(self, including_in_progress=False):
         """ Small helper to get the currently active consolidated invoice if more that one is linked to an order. """
-        return self.env['myinvois.document'].union(*[order.consolidated_invoice_ids.filtered(lambda i: i.myinvois_state != 'cancelled')[:1] for order in self])
+        return self.env['myinvois.document'].union(*[order.consolidated_invoice_ids._get_active_myinvois_document(including_in_progress) for order in self])
