+++ addons/l10n_my_edi_pos/security/ir.model.access.csv
@@ -1,4 +1,2 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_myinvois_document_user,access_myinvois_document_user,model_myinvois_document,account.group_account_invoice,1,1,1,1
 access_myinvois_consolidate_invoice_wizard_user,access_myinvois_consolidate_invoice_wizard_user,model_myinvois_consolidate_invoice_wizard,account.group_account_invoice,1,1,1,1
-access_myinvois_document_status_update_wizard_user,access_myinvois_document_status_update_wizard_user,model_myinvois_document_status_update_wizard,account.group_account_invoice,1,1,1,1
