+++ addons/l10n_my_edi_pos/tests/__init__.py
@@ -1,3 +1,2 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from . import test_myinvois_pos
