+++ addons/l10n_my_edi_pos/tests/expected_xmls/consolidated_invoice.xml
@@ -7,6 +7,7 @@
     <cbc:DocumentCurrencyCode>USD</cbc:DocumentCurrencyCode>
     <cac:AdditionalDocumentReference>
         <cbc:ID>123456789</cbc:ID>
+        <cbc:DocumentType>K2</cbc:DocumentType>
     </cac:AdditionalDocumentReference>
     <cac:AccountingSupplierParty>
         <cac:Party>
@@ -36,48 +37,8 @@
                     <cbc:Name>Malaysia</cbc:Name>
                 </cac:Country>
             </cac:PostalAddress>
-            <cac:PartyTaxScheme>
-                <cbc:RegistrationName>MY Test Company</cbc:RegistrationName>
-                <cbc:CompanyID>C2584563200</cbc:CompanyID>
-                <cac:RegistrationAddress>
-                    <cbc:CityName>Kuala Lumpur</cbc:CityName>
-                    <cbc:PostalZone>50300</cbc:PostalZone>
-                    <cbc:CountrySubentity>Kuala Lumpur</cbc:CountrySubentity>
-                    <cbc:CountrySubentityCode>14</cbc:CountrySubentityCode>
-                    <cac:AddressLine>
-                        <cbc:Line>1 Wisma Dato Dagang</cbc:Line>
-                    </cac:AddressLine>
-                    <cac:AddressLine>
-                        <cbc:Line>Jln Raja Alang Kampung Bahru Mala</cbc:Line>
-                    </cac:AddressLine>
-                    <cac:Country>
-                        <cbc:IdentificationCode listID="ISO3166-1" listAgencyID="6">MYS</cbc:IdentificationCode>
-                        <cbc:Name>Malaysia</cbc:Name>
-                    </cac:Country>
-                </cac:RegistrationAddress>
-                <cac:TaxScheme>
-                    <cbc:ID>VAT</cbc:ID>
-                </cac:TaxScheme>
-            </cac:PartyTaxScheme>
             <cac:PartyLegalEntity>
                 <cbc:RegistrationName>MY Test Company</cbc:RegistrationName>
-                <cbc:CompanyID>C2584563200</cbc:CompanyID>
-                <cac:RegistrationAddress>
-                    <cbc:CityName>Kuala Lumpur</cbc:CityName>
-                    <cbc:PostalZone>50300</cbc:PostalZone>
-                    <cbc:CountrySubentity>Kuala Lumpur</cbc:CountrySubentity>
-                    <cbc:CountrySubentityCode>14</cbc:CountrySubentityCode>
-                    <cac:AddressLine>
-                        <cbc:Line>1 Wisma Dato Dagang</cbc:Line>
-                    </cac:AddressLine>
-                    <cac:AddressLine>
-                        <cbc:Line>Jln Raja Alang Kampung Bahru Mala</cbc:Line>
-                    </cac:AddressLine>
-                    <cac:Country>
-                        <cbc:IdentificationCode listID="ISO3166-1" listAgencyID="6">MYS</cbc:IdentificationCode>
-                        <cbc:Name>Malaysia</cbc:Name>
-                    </cac:Country>
-                </cac:RegistrationAddress>
             </cac:PartyLegalEntity>
             <cac:Contact>
                 <cbc:ID>___ignore___</cbc:ID>
@@ -101,7 +62,7 @@
             <cac:PostalAddress>
                 <cbc:CityName>NA</cbc:CityName>
                 <cbc:CountrySubentity>Johor</cbc:CountrySubentity>
-                <cbc:CountrySubentityCode>01</cbc:CountrySubentityCode>
+                <cbc:CountrySubentityCode>17</cbc:CountrySubentityCode>
                 <cac:AddressLine>
                     <cbc:Line>NA</cbc:Line>
                 </cac:AddressLine>
@@ -110,40 +71,8 @@
                     <cbc:Name>Malaysia</cbc:Name>
                 </cac:Country>
             </cac:PostalAddress>
-            <cac:PartyTaxScheme>
-                <cbc:RegistrationName>General Public</cbc:RegistrationName>
-                <cbc:CompanyID>EI00000000010</cbc:CompanyID>
-                <cac:RegistrationAddress>
-                    <cbc:CityName>NA</cbc:CityName>
-                    <cbc:CountrySubentity>Johor</cbc:CountrySubentity>
-                    <cbc:CountrySubentityCode>01</cbc:CountrySubentityCode>
-                    <cac:AddressLine>
-                        <cbc:Line>NA</cbc:Line>
-                    </cac:AddressLine>
-                    <cac:Country>
-                        <cbc:IdentificationCode listID="ISO3166-1" listAgencyID="6">MYS</cbc:IdentificationCode>
-                        <cbc:Name>Malaysia</cbc:Name>
-                    </cac:Country>
-                </cac:RegistrationAddress>
-                <cac:TaxScheme>
-                    <cbc:ID>VAT</cbc:ID>
-                </cac:TaxScheme>
-            </cac:PartyTaxScheme>
             <cac:PartyLegalEntity>
                 <cbc:RegistrationName>General Public</cbc:RegistrationName>
-                <cbc:CompanyID>EI00000000010</cbc:CompanyID>
-                <cac:RegistrationAddress>
-                    <cbc:CityName>NA</cbc:CityName>
-                    <cbc:CountrySubentity>Johor</cbc:CountrySubentity>
-                    <cbc:CountrySubentityCode>01</cbc:CountrySubentityCode>
-                    <cac:AddressLine>
-                        <cbc:Line>NA</cbc:Line>
-                    </cac:AddressLine>
-                    <cac:Country>
-                        <cbc:IdentificationCode listID="ISO3166-1" listAgencyID="6">MYS</cbc:IdentificationCode>
-                        <cbc:Name>Malaysia</cbc:Name>
-                    </cac:Country>
-                </cac:RegistrationAddress>
             </cac:PartyLegalEntity>
             <cac:Contact>
                 <cbc:ID>___ignore___</cbc:ID>
