+++ addons/l10n_my_edi_pos/tests/expected_xmls/consolidated_invoice_refund.xml
@@ -108,6 +108,9 @@
             </cac:PartyLegalEntity>
         </cac:DeliveryParty>
     </cac:Delivery>
+    <cac:PrepaidPayment>
+        <cbc:PaidAmount currencyID="USD">420.00</cbc:PaidAmount>
+    </cac:PrepaidPayment>
     <cac:TaxExchangeRate>
         <cbc:SourceCurrencyCode>USD</cbc:SourceCurrencyCode>
         <cbc:TargetCurrencyCode>MYR</cbc:TargetCurrencyCode>
@@ -132,7 +135,7 @@
         <cbc:LineExtensionAmount currencyID="USD">400.00</cbc:LineExtensionAmount>
         <cbc:TaxExclusiveAmount currencyID="USD">400.00</cbc:TaxExclusiveAmount>
         <cbc:TaxInclusiveAmount currencyID="USD">420.00</cbc:TaxInclusiveAmount>
-        <cbc:PrepaidAmount currencyID="USD">420.00</cbc:PrepaidAmount>
+        <cbc:PrepaidAmount currencyID="USD">0.00</cbc:PrepaidAmount>
         <cbc:PayableAmount currencyID="USD">0.00</cbc:PayableAmount>
     </cac:LegalMonetaryTotal>
     <cac:InvoiceLine>
