--- addons/l10n_my_edi_pos/views/account_tax_view.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <record id="view_tax_form_inherit_l10n_my_edi_pos" model="ir.ui.view">
-        <field name="name">account.tax.form.inherit.l10n_my_edi_pos</field>
-        <field name="model">account.tax</field>
-        <field name="inherit_id" ref="l10n_my_edi.view_tax_form_inherit_l10n_my_myinvois"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='l10n_my_tax_type']" position="after">
-                <field name="l10n_my_tax_exemption_reason" invisible="l10n_my_tax_type != 'E'"/>
-            </xpath>
-        </field>
-    </record>
-</odoo>
