+++ addons/l10n_my_edi_pos/views/myinvois_document_pos_views.xml
@@ -4,7 +4,7 @@
         <record id="myinvois_document_pos_form_view" model="ir.ui.view">
             <field name="name">myinvois.document.pos.form.view</field>
             <field name="model">myinvois.document</field>
-            <field name="inherit_id" ref="l10n_my_edi_pos.myinvois_document_form_view"/>
+            <field name="inherit_id" ref="l10n_my_edi.myinvois_document_form_view"/>
             <field name="mode">primary</field>
             <field name="arch" type="xml">
                 <div name="button_box" position="inside">
@@ -23,7 +23,7 @@
         <record id="myinvois_document_pos_list_view" model="ir.ui.view">
             <field name="name">myinvois.document.list.view</field>
             <field name="model">myinvois.document</field>
-            <field name="inherit_id" ref="l10n_my_edi_pos.myinvois_document_list_view"/>
+            <field name="inherit_id" ref="l10n_my_edi.myinvois_document_list_view"/>
             <field name="mode">primary</field>
             <field name="arch" type="xml">
                 <header position="inside">
