+++ addons/l10n_my_edi_pos/views/pos_order_views.xml
@@ -7,7 +7,7 @@
         <field name="arch" type="xml">
             <xpath expr="//div[@name='button_box']" position="inside">
                 <field name="consolidated_invoice_ids" invisible="1"/>
-                <button name="action_show_consolidated_invoice"
+                <button name="action_show_myinvois_documents"
                         type="object"
                         icon="fa-sun-o"
                         class="oe_stat_button"
