+++ addons/l10n_my_edi_pos/wizard/__init__.py
@@ -1,4 +1,3 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import myinvois_consolidate_invoice_wizard
-from . import myinvois_status_update_wizard
